feat: wire settings routes — /settings/* sub-routing with lazy-loaded category pages
Replace monolithic SettingsPage with a Routes-based router that maps
/settings/{category} to individual settings pages. Old settings page
preserved as LegacySettings.tsx for type/constant re-exports.

- /settings → SettingsOverview (card grid)
- /settings/general, connections, subtitles, providers, automation,
  translation, notifications, system → respective category pages
- Redirect /statistics, /tasks, /plugins, /logs → settings sub-routes
- Remove unused lazy imports from App.tsx
- Fix test mocks for renamed/removed pages
