+++ frontend/src/test/App.skiplink.test.tsx
@@ -49,18 +49,10 @@ vi.mock('react-i18next', () => ({
 
 // Mock all lazy-loaded pages to avoid dynamic import issues in tests
 vi.mock('@/pages/Dashboard', () => ({ Dashboard: () => <div>Dashboard</div> }))
-vi.mock('@/pages/Activity', () => ({ ActivityPage: () => <div>Activity</div> }))
-vi.mock('@/pages/Wanted', () => ({ WantedPage: () => <div>Wanted</div> }))
-vi.mock('@/pages/Queue', () => ({ QueuePage: () => <div>Queue</div> }))
+vi.mock('@/pages/ActivityPage', () => ({ ActivityPage: () => <div>Activity</div> }))
 vi.mock('@/pages/Settings', () => ({ SettingsPage: () => <div>Settings</div> }))
-vi.mock('@/pages/Logs', () => ({ LogsPage: () => <div>Logs</div> }))
-vi.mock('@/pages/Statistics', () => ({ StatisticsPage: () => <div>Statistics</div> }))
 vi.mock('@/pages/Library', () => ({ LibraryPage: () => <div>Library</div> }))
 vi.mock('@/pages/SeriesDetail', () => ({ SeriesDetailPage: () => <div>SeriesDetail</div> }))
-vi.mock('@/pages/History', () => ({ HistoryPage: () => <div>History</div> }))
-vi.mock('@/pages/Blacklist', () => ({ BlacklistPage: () => <div>Blacklist</div> }))
-vi.mock('@/pages/Tasks', () => ({ TasksPage: () => <div>Tasks</div> }))
-vi.mock('@/pages/Plugins', () => ({ PluginsPage: () => <div>Plugins</div> }))
 vi.mock('@/pages/NotFound', () => ({ NotFoundPage: () => <div>NotFound</div> }))
 vi.mock('@/pages/Onboarding', () => ({ default: () => <div>Onboarding</div> }))
